refactor(docs): consolidate Mermaid into a single plugins/mermaid/ folder
Align the main docs/ site with libs/Momentum/docs/ structure: merge the
markdown-it fence transform (plugins/mermaid.ts) and the runtime
render/init (theme/mermaid.ts) into one self-contained
plugins/mermaid/mermaid.ts, and move Mermaid.vue alongside it.

- mermaid + ELK are now lazy-imported inside getMermaid() (matches libs,
  avoids eager browser-API imports during SSR)
- add vite.ssr.noExternal: ['mermaid'] so SSR bundles mermaid correctly
- update config.mts and theme/index.ts to the new paths

Verified with an isolated VitePress build: the fence transform fires and
a mermaid page renders through SSR without error.
